LIBSEARCH-53. Initial implementation of JSON-LD parse filter
Initial implementation of a JSON-LD parse filter plugin for Apache
Nutch.

When parsing web pages, the HTML is searched for JSON-LD scripts,
and if one of the script contains a "WebPage" object, the "text" value
of the object is placed in the "documentContent" field of the parse
content metadata.

Note: This plugin is not enabled by default, and no Nutch configuration
files have been changed. See the docker_config/nutch/ subdirectory in
the "searchumd" application for the relevant configuration files.

https://issues.umd.edu/browse/LIBSEARCH-53